Fix bug #279 spreads are not properties.
The bug is back because we were only checking for CallExpressions, if you add two Conditionals it will fail again.
It should be enough to disregard rests and spreads all together, they shouldn't actually be in the properties list anyway.
This fixes the bug, test data modified to reproduce, with no new tests failing so we have to assume it is good.